Refactor : useSchedules.js로 분리
