+++ src/pages/calendar/components/CalendarSchedule.jsx
@@ -1,7 +1,5 @@
-import { useEffect, useState } from "react";
 import styled from "styled-components";
-import { db, auth } from "../../../shared/firebase";
-import { collection, onSnapshot, query, where } from "firebase/firestore";
+import useSchedules from "../hooks/useSchedules";
 
 const StyledCalendarDate = styled.tbody`
   tr {
@@ -71,36 +69,7 @@ const ScheduleBar = styled.span`
 `
 
 const CalendarSchedule = ({ weeks, handleScheduleClick }) => {
-  const [schedules, setSchedules] = useState([]);
-
-  useEffect(() => {
-    const user = auth.currentUser;
-    if (!user) return;
-  
-    // 로그인한 유저의 일정만 가져옴
-    const schedulesQuery = query(collection(db, "schedules"),
-      where("userId", "==", user.uid)
-    );
-  
-    const unsubscribe = onSnapshot(schedulesQuery, (snapshot) => {
-      const schedules = snapshot.docs.map((doc) => {
-        const { userId, title, startDate, endDate, selectedColor, contents } = doc.data();
-        return {
-          userId,
-          title,
-          startDate: startDate.toDate(),
-          endDate: endDate.toDate(),
-          selectedColor,
-          contents,
-          id: doc.id,
-        };
-      });
-      schedules.sort((a, b) => a.startDate - b.startDate);
-
-      setSchedules(schedules);
-    });
-    return () => unsubscribe();
-  }, [])
+  const schedules = useSchedules();
 
   // 날짜에서 연, 월, 일 정보만 남기고 시간 제거
   const getDateOnly = (date) => new Date(date.getFullYear(), date.getMonth(), date.getDate());
